rename stops file
